requirements.txt
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
pywin32==308
pywin32-ctypes==0.2.3
python-dotenv==1.0.1
python-telegram-bot==21.6
requests
httpx==0.27
